BSC/BSCScanAPI/pyproject.toml
[tool.poetry]
name = "bscscanapi"
version = "0.1.0"
description = ""
authors = ["cardene"]

[tool.poetry.dependencies]
python = "^3.9"
bscscan-python = "^2.0.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
